Move PrefixResolver, PrefixResolverSyntax to logic.attribute
