+++ src/main/java/seedu/address/logic/autocomplete/AutoCompletePrefixResolver.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import seedu.address.logic.attribute.PrefixResolver;
 import seedu.address.logic.parser.Prefix;
 
 /**
